+++ changelog.md
@@ -32,6 +32,7 @@ Codion Change Log
 - ColumnConditionPanel, add/removeListener methods replaced with methods that return the underlying event observer.
 - FilteredTable, add/removeListener methods replaced with methods that return the underlying event observer.
 - FilteredTableConditionPanel, add/removeFocusGainedListener() methods replaced focusGainedObserver() method.
+- NumberField.addListener() replaced with numberObserver() method.
 ### is.codion.swing.framework.ui
 - EntityPanel, add/removeListener methods replaced with methods that return the underlying event observer.